refactor(ui): Cycle 2 frontend — type safety, accessibility, validation, error UX
Across all 3 apps (memberry, admin, account):
- Eliminate unjustified as-any casts (105→0), shared ApiListResponse/ApiErrorBody types
- Centralize zodResolver Zod v4 wrapper, add no-explicit-any ESLint (warn)
- Replace 21 raw HTML with @monobase/ui, add role="alert"/aria-live to 51 error states
- Wire react-hook-form+zod into all forms, add ErrorState component with retry
- Fix tsc errors: bigint registrationFee, CheckInByPersonBody interface
- ESLint: no-restricted-syntax for raw HTML (error), test file overrides

+++ apps/memberry/src/features/events/components/event-form.tsx
@@ -113,7 +113,7 @@ export function EventForm({ orgId, event, onSuccess, onCancel }: EventFormProps)
       startDate: new Date(data.startDate),
       endDate: new Date(data.endDate),
       location: data.location || undefined,
-      registrationFee: Math.round((data.registrationFee ?? 0) * 100),
+      registrationFee: BigInt(Math.round((data.registrationFee ?? 0) * 100)),
       capacity: data.capacity && !Number.isNaN(data.capacity) ? data.capacity : undefined,
       creditBearing: false,
     }
